% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_continuous_nwt.R
\name{weight_continuous_nwt}
\alias{weight_continuous_nwt}
\title{Weight for the continuous effect sizes}
\usage{
weight_continuous_nwt(alpha, et, ranksProb, x0 = NULL)
}
\arguments{
\item{alpha}{Numeric, significance level of the hypothesis test}

\item{et}{Numeric, mean effect size of the test statistics}

\item{ranksProb}{Numeric vector of ranks probability of the tests given
the effect size}

\item{x0}{Numeric, a initial value for the Newton-Raphson mehod.}
}
\value{
\code{w} Numeric vector of weights of the tests for the continuous case

\code{lambda} Numeric value of the LaGrange multiplier

\code{n} Integer, number of iteration needed to obtain the initial value

\code{k} Integer, number of iteration needed to obtain the optimal
value of \code{lambda}
}
\description{
This is a Newton Raphson based algorithm to compute weight from
the ranks probability for the continuous effect sizes.
}
\details{
If one wants to test \deqn{H_0: epsilon_i = 0 vs. H_a: \epsilon_i > 0,}
then \code{et} should be mean of the test effect sizes.
This is called hypothesis testing for the continuous effect sizes.\cr

For the Newton-Raphson mehthod the initial value \code{x0} is very important.
One may need to supply externally if the function shows error. Newton-Raphson
method may not work in some cases. In that situations, use \code{weight_continuous}
function, which is a general but slow approach to compute weights beased on the
grid search algorithm.
}
\examples{

# compute the probabilities of the ranks of a test being rank 1 to 100 if the
# targeted test effect is 2 and the overall mean covariate effect is 1.
ranks <- 1:100
prob <- sapply(ranks, prob_rank_givenEffect, et = 2, ey = 1, nrep = 10000,
                              m0 = 50, m1 = 50)

# compute weight for the continuous case
results = weight_continuous_nwt(alpha = .05, et = 2,
                                             ranksProb = prob)

}
\seealso{
\code{\link{prob_rank_givenEffect}} \code{\link{weight_continuous}}
}
\author{
Mohamad S. Hasan, shakilmohamad7@gmail.com
}
